requirements.txt
# Dependencias del Sistema de Gestión IRC

# Google Sheets API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Procesamiento de datos
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9

# Procesamiento de PDFs
pdfplumber==0.10.3
PyPDF2==3.0.1
pdfrw==0.4
reportlab==4.0.7

# Gráficos y visualización
matplotlib==3.8.2
Pillow==10.1.0

# GUI
tkcalendar==1.6.1

# Utilidades
python-dateutil==2.8.2
pytz==2023.3.post1

# Logging y debugging
colorlog==6.8.0
